extract cyclically referenced constants
