+++ src/models/committee.ts
@@ -1,7 +1,8 @@
 import {MemberData, MemberID, MemberOption, membersToOptions, membersToPresentOptions, Rank} from '../modules/member';
 import {logCreateMember} from '../modules/analytics';
 import _ from 'lodash';
-import {CaucusData, CaucusID, DEFAULT_CAUCUS, DEFAULT_CAUCUS_TIME_SECONDS} from "./caucus";
+import {CaucusData, CaucusID, DEFAULT_CAUCUS} from "./caucus";
+import { DEFAULT_CAUCUS_TIME_SECONDS } from './constants';
 import firebase from "firebase/compat/app";
 import {PostData, PostID} from "./post";
 import {MotionData, MotionID} from "./motion";
